rag-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
httpx
qdrant-client
psycopg2-binary
asyncpg
python-multipart
requests
